Play list solo para el usuario que inicio sesion

+++ src/main/java/com/espotify/mysql/service/PlayListServiceImpl.java
@@ -1,11 +1,15 @@
 package com.espotify.mysql.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.espotify.mysql.model.PlayList;
+import com.espotify.mysql.model.User;
 import com.espotify.mysql.repository.PlayListRepository;
 
 @Service
@@ -15,8 +19,13 @@ public class PlayListServiceImpl implements PlayListService {
 	private PlayListRepository playListRepository;
 
 	@Override
-	public PlayList getPlayListById(Integer playListId) {
-		return playListRepository.findById(playListId).orElse(null);
+	public PlayList getPlayListById(User user, Integer playListId) {
+		PlayList playList = playListRepository.findById(playListId).orElse(null);
+
+		if (playList != null && playList.getUser().equals(user))
+			return playList;
+
+		return null;
 	}
 
 	@Override
@@ -25,18 +34,31 @@ public class PlayListServiceImpl implements PlayListService {
 	}
 
 	@Override
-	public void deletePlayList(PlayList playList) {
-		playListRepository.delete(playList);
+	public void deletePlayList(User user, PlayList playList) {
+		if (playList.getUser().equals(user))
+			playListRepository.delete(playList);
 	}
 
 	@Override
-	public List<PlayList> getAllPlayLists() {
-		return playListRepository.findAll();
+	public List<PlayList> getAllPlayLists(User user) {
+		List<PlayList> result = new ArrayList<>();
+		List<PlayList> playLists = playListRepository.findAll();
+
+		playLists.forEach(playList -> { 
+			if (playList.getUser().equals(user)) {
+				result.add(playList);
+			}
+		});
+
+		return result;
 	}
 
 	@Override
-	public void deletePlayListById(Integer playListId) {
-		playListRepository.deleteById(playListId);
+	public void deletePlayListById(User user, Integer playListId) {
+		PlayList playList = getPlayListById(user, playListId);
+
+		if (playList != null)
+			playListRepository.deleteById(playListId);
 	}
 	
 }
